process_sonlandir() eksiklikler giderildi.

+++ procx.c
@@ -294,7 +294,6 @@ void process_baslat(const char *command, ProcessMode mode){ //command: kullanici
     else{
         //burada kritik bölge mevzuları olacak!!!
         // PARENT PROCESS
-        sem_wait(g_sem);
         //shared memory'ye process kaydi ekle
         add_process_record(child_pid, command, mode);
 
@@ -404,14 +403,23 @@ void process_sonlandir(){
 
     sem_wait(g_sem);
 
+    int found = 0;
     //active ve pid target_pid ye eşitse kill yap ve bilidim gönder
     for(int i = 0; i<50; i++){
-        if(g_shared->processes[i].is_active && g_shared->processes[i].pid == target_pid){
+        if(g_shared->processes[i].is_active && g_shared->processes[i].pid == target_pid && 
+            g_shared->processes[i].owner_pid == getpid()){
             kill(target_pid, SIGTERM);
-        }
-        send_notification(2, target_pid);
+            found = 1;
+            break;
+        }  
     }
     sem_post(g_sem);
+    if(found){
+        send_notification(2, target_pid);
+    }else{
+        printf("Bu PID'e ait aktif process bulunamadi. \n");
+    }
+    
 }
 
 void display_menu() {
